assignment-3-784915/-/code_Apache_(NOT_FUNCTIONAL)/requirements.txt
cassandra-driver
numpy
pandas
argparse
matplotlib
tqdm
kafka-python